some more stuff

+++ client/client.lua
@@ -5,11 +5,10 @@ local bringingVehicle = false
 local hasAlreadyEnteredMarker = false
 local randomLocation
 local LastZone
-local blip
 local location = nil
-local alreadyDone --variable to only let players complete the quest once per server reboot (optimal when daily)
-local todaysPrimeModel = "car I guess"
-local todaysPrimeColor = "unidentified"
+local alreadyDone -- variable to only let players complete the quest once per server reboot (optimal when daily)
+local todaysPrimeModel = '';
+local todaysPrimeColor = '';
 
 Citizen.CreateThread(function()
 	while ESX == nil do
@@ -43,20 +42,20 @@ local function capitalizeFirstLetter(word)
     return (string.upper(firstLetter) .. string.lower(restOfTheWord))
 end
 
+local activeBlip = nil;
 AddEventHandler('carhunting:locationandblip', function()
-	if DoesBlipExist(blip) then
-		RemoveBlip(blip)
-	end
+	if DoesBlipExist(activeBlip) then RemoveBlip(activeBlip) end
 
 	if not alreadyDone or Config.Repeatable then
-		blip = AddBlipForCoord(location.NPCcoords)
-		SetBlipSprite(blip, 530)
-		SetBlipAsShortRange(blip, true)
-		SetBlipDisplay(blip, 2)
-		SetBlipColour (blip, 23)
+		activeBlip = AddBlipForCoord(location.NPCcoords)
+		SetBlipSprite(activeBlip, 530)
+		SetBlipAsShortRange(activeBlip, true)
+		SetBlipDisplay(activeBlip, 2)
+		SetBlipColour (activeBlip, 23)
+
 		BeginTextCommandSetBlipName('STRING')
 		AddTextComponentSubstringPlayerName(_U('blip_name'))
-		EndTextCommandSetBlipName(blip)
+		EndTextCommandSetBlipName(activeBlip)
 	end
 end)
 
@@ -76,7 +75,7 @@ AddEventHandler('carhunting:hasEnteredMarker', function(zone)
             modelName = modelName,
             Plate = ESX.Math.Trim(GetVehicleNumberPlateText(currentVehicle)),
             bodyHealth = GetVehicleBodyHealth(currentVehicle)
-        }
+        } -- We store this to the table so we don't have to take the info again next time.
     else
         ESX.ShowHelpNotification(_U('bring_car')); bringingVehicle = false;
     end
@@ -85,6 +84,7 @@ end)
 AddEventHandler('carhunting:hasExitedMarker', function(zone) bringingVehicle = false; currentVehicleData = {}; end)
 
 
+local pedModel = 'a_m_m_og_boss_01';
 function StartThreads()
     -- Enter / Exit marker events & Draw Markers
     Citizen.CreateThread(function()
@@ -98,21 +98,14 @@ function StartThreads()
 
                 if distance < 40 then
                     letSleep = false
+
                     DrawMarker(1, location.SellPoint, 0.0, 0.0, 0.0, location.Rotation.x, location.Rotation.y, location.Rotation.z, 2.5, 2.5, 1.0, 255, 0, 0, 100, false, false, 2, false, nil, nil, false)
 
-
-                    --spawn NPC if not yet
                     if not npcSpawned then
-                        RequestModel(GetHashKey("a_m_m_og_boss_01"))
-        
-                        while not HasModelLoaded(GetHashKey("a_m_m_og_boss_01")) do
-                            Wait(1)
-                        end
-                        npc = CreatePed(4, "a_m_m_og_boss_01", location.NPCcoords, location.NPCheading, false, true)
-                        FreezeEntityPosition(npc, true)
-                        SetEntityInvincible(npc, true)
-                        SetBlockingOfNonTemporaryEvents(npc, true)
-                        npcSpawned = true
+                        ESX.Streaming.RequestModel(pedModel, function()
+                            npc = CreatePed(4, pedModel, location.NPCcoords, location.NPCheading, false, true); npcSpawned = true;
+                            FreezeEntityPosition(npc, true); SetEntityInvincible(npc, true); SetBlockingOfNonTemporaryEvents(npc, true)
+                        end)
                     end
 
                     if distance < 2 then
@@ -125,8 +118,6 @@ function StartThreads()
                 if distanceToTalk <= 2 then
                     ESX.ShowHelpNotification(_U('request', string.lower(todaysPrimeColor), capitalizeFirstLetter(todaysPrimeModel))) -- for some languages you might want to switch the 2 variables' order
                 end
-
-
             end
 
             if (isInMarker and not HasAlreadyEnteredMarker) or (isInMarker and LastZone ~= currentZone) then
@@ -140,30 +131,23 @@ function StartThreads()
             end
 
             if letSleep or (alreadyDone and not Config.Repeatable) then
-                npcSpawned = false
-                if npc then --delete npc when we're away
-                    DeletePed(npc)
-                end
-                Citizen.Wait(500)
+                if npc ~= nil then DeletePed(npc); npcSpawned = false; end; Wait(500)
             end
         end
     end)
 
-    -- Key controls
     Citizen.CreateThread(function()
         while true do
 
             if bringingVehicle then
-                local playerPed = PlayerPedId();
-                local currentVehicle = GetVehiclePedIsIn(playerPed, false);
-
+                local playerPed = PlayerPedId(); local currentVehicle = GetVehiclePedIsIn(playerPed, false);
                 if DoesEntityExist(currentVehicle) and GetPedInVehicleSeat(currentVehicle, -1) == playerPed then
                     if IsControlJustReleased(0, 38) and not isDead then
                         TriggerServerEvent('carhunting:submitvehicle', currentVehicleData)
                     end
                 end
             else
-                Wait(500)
+                Wait(500) -- This loop won't have to run at 0 ms when not needed.
             end
 
             Wait(0)
@@ -171,31 +155,34 @@ function StartThreads()
     end)
 end
 
+local totalTimeout = 100;
 RegisterNetEvent('carhunting:accepted')
 AddEventHandler('carhunting:accepted', function(reward)
 	local playerPed = PlayerPedId()
 	local currentVehicle = GetVehiclePedIsIn(playerPed, false);
 
-	ESX.Game.DeleteVehicle(currentVehicle);
+    if not DoesEntityExist(currentVehicle) then return end
+    
+    ESX.Game.DeleteVehicle(currentVehicle);
 
-	if reward == Config.FullReward then
-		ESX.ShowNotification(_U('sold_for', reward))
-		ESX.ShowHelpNotification(_U('great_work'))
-	else
-		ESX.ShowNotification(_U('sold_for', reward))
-		ESX.ShowHelpNotification(_U('job_done'))
-	end
+    if reward == Config.FullReward then
+        ESX.ShowNotification(_U('sold_for', reward))
+        ESX.ShowHelpNotification(_U('great_work'))
+    else
+        ESX.ShowNotification(_U('sold_for', reward))
+        ESX.ShowHelpNotification(_U('job_done'))
+    end
 
-	ESX.TriggerServerCallback('carhunting:picktodaysvehicle', function(primeColor, primeModel, spawnNumber, cleared)
-		todaysPrimeColor = primeColor
-		todaysPrimeModel = primeModel
-		location = Config.Coords[spawnNumber]
-		alreadyDone = cleared
-	end)
+    ESX.TriggerServerCallback('carhunting:picktodaysvehicle', function(Data)
+        todaysPrimeColor = Data.primeColor
+        todaysPrimeModel = Data.primeModel
+        location = Config.Coords[Data.spawnNumber]
+        alreadyDone = Data.hasCompleted
+    end)
     
     Wait(500);
     
-	TriggerEvent('carhunting:locationandblip')
+    TriggerEvent('carhunting:locationandblip')
 end)
